Move all but the apply tests
