+++ python/cudf/cudf/tests/groupby/test_shift.py
@@ -6,7 +6,7 @@ import pandas as pd
 import pytest
 
 import cudf
-from cudf.testing import assert_groupby_results_equal
+from cudf.testing import assert_eq, assert_groupby_results_equal
 from cudf.testing.dataset_generator import rand_dataframe
 
 
@@ -214,3 +214,13 @@ def test_groupby_select_then_shift():
     actual = gdf.groupby("a")["c"].shift(1)
 
     assert_groupby_results_equal(expected, actual)
+
+
+def test_groupby_shift_series_multiindex():
+    idx = cudf.MultiIndex.from_tuples(
+        [("a", 1), ("a", 2), ("b", 1), ("b", 2)], names=["f", "s"]
+    )
+    ser = cudf.Series(range(4), index=idx)
+    result = ser.groupby(level=0).shift(1)
+    expected = ser.to_pandas().groupby(level=0).shift(1)
+    assert_eq(expected, result)
