+++ python/cudf/cudf/tests/test_groupby.py
@@ -8,7 +8,7 @@ import pandas as pd
 import pytest
 
 import cudf
-from cudf import DataFrame, Series
+from cudf import DataFrame
 from cudf.core._compat import (
     PANDAS_CURRENT_SUPPORTED_VERSION,
     PANDAS_GE_220,
@@ -955,297 +955,3 @@ def test_group_by_empty_apply(request, dtype, apply_op):
         check_dtype=True,
         check_index_type=True,
     )
-
-
-def test_groupby_consecutive_operations():
-    df = cudf.DataFrame([[1, np.nan], [1, 4], [5, 6]], columns=["A", "B"])
-    pdf = df.to_pandas()
-
-    gg = df.groupby("A")
-    pg = pdf.groupby("A")
-
-    actual = gg.nth(-1)
-    expected = pg.nth(-1)
-
-    assert_groupby_results_equal(actual, expected, check_dtype=False)
-
-    actual = gg.nth(0)
-    expected = pg.nth(0)
-
-    assert_groupby_results_equal(actual, expected, check_dtype=False)
-
-    actual = gg.cumsum()
-    expected = pg.cumsum()
-
-    assert_groupby_results_equal(actual, expected, check_dtype=False)
-
-    actual = gg.cumcount()
-    expected = pg.cumcount()
-
-    assert_groupby_results_equal(actual, expected, check_dtype=False)
-
-    actual = gg.cumsum()
-    expected = pg.cumsum()
-
-    assert_groupby_results_equal(actual, expected, check_dtype=False)
-
-
-@pytest.mark.skipif(
-    PANDAS_VERSION < PANDAS_CURRENT_SUPPORTED_VERSION,
-    reason="Warning only given on newer versions.",
-)
-def test_categorical_grouping_pandas_compatibility():
-    gdf = cudf.DataFrame(
-        {
-            "key": cudf.Series([2, 1, 3, 1, 1], dtype="category"),
-            "a": [0, 1, 3, 2, 3],
-        }
-    )
-    pdf = gdf.to_pandas()
-
-    with cudf.option_context("mode.pandas_compatible", True):
-        actual = gdf.groupby("key", sort=False).sum()
-    with pytest.warns(FutureWarning):
-        # observed param deprecation.
-        expected = pdf.groupby("key", sort=False).sum()
-    assert_eq(actual, expected)
-
-
-@pytest.mark.parametrize("normalize", [True, False])
-@pytest.mark.parametrize("sort", [True, False])
-@pytest.mark.parametrize("ascending", [True, False])
-@pytest.mark.parametrize("dropna", [True, False])
-@pytest.mark.parametrize("as_index", [True, False])
-def test_group_by_value_counts(normalize, sort, ascending, dropna, as_index):
-    # From Issue#12789
-    df = cudf.DataFrame(
-        {
-            "gender": ["male", "male", "female", "male", "female", "male"],
-            "education": ["low", "medium", np.nan, "low", "high", "low"],
-            "country": ["US", "FR", "US", "FR", "FR", "FR"],
-        }
-    )
-    pdf = df.to_pandas()
-
-    actual = df.groupby("gender", as_index=as_index).value_counts(
-        normalize=normalize, sort=sort, ascending=ascending, dropna=dropna
-    )
-    expected = pdf.groupby("gender", as_index=as_index).value_counts(
-        normalize=normalize, sort=sort, ascending=ascending, dropna=dropna
-    )
-
-    # TODO: Remove `check_names=False` once testing against `pandas>=2.0.0`
-    assert_groupby_results_equal(
-        actual,
-        expected,
-        check_index_type=False,
-        as_index=as_index,
-        by=["gender", "education"],
-        sort=sort,
-    )
-
-
-def test_group_by_value_counts_subset():
-    # From Issue#12789
-    df = cudf.DataFrame(
-        {
-            "gender": ["male", "male", "female", "male", "female", "male"],
-            "education": ["low", "medium", "high", "low", "high", "low"],
-            "country": ["US", "FR", "US", "FR", "FR", "FR"],
-        }
-    )
-    pdf = df.to_pandas()
-
-    actual = df.groupby("gender").value_counts(["education"])
-    expected = pdf.groupby("gender").value_counts(["education"])
-
-    # TODO: Remove `check_names=False` once testing against `pandas>=2.0.0`
-    assert_groupby_results_equal(
-        actual, expected, check_names=False, check_index_type=False
-    )
-
-
-def test_group_by_value_counts_clash_with_subset():
-    df = cudf.DataFrame({"a": [1, 5, 3], "b": [2, 5, 2]})
-    with pytest.raises(ValueError):
-        df.groupby("a").value_counts(["a"])
-
-
-def test_group_by_value_counts_subset_not_exists():
-    df = cudf.DataFrame({"a": [1, 5, 3], "b": [2, 5, 2]})
-    with pytest.raises(ValueError):
-        df.groupby("a").value_counts(["c"])
-
-
-def test_group_by_value_counts_with_count_column():
-    df = cudf.DataFrame({"a": [1, 5, 3], "count": [2, 5, 2]})
-    with pytest.raises(ValueError):
-        df.groupby("a", as_index=False).value_counts()
-
-
-def test_groupby_internal_groups_empty(gdf):
-    # test that we don't segfault when calling the internal
-    # .groups() method with an empty list:
-    gb = gdf.groupby("y")
-    _, _, grouped_vals = gb._groups([])
-    assert grouped_vals == []
-
-
-def test_groupby_shift_series_multiindex():
-    idx = cudf.MultiIndex.from_tuples(
-        [("a", 1), ("a", 2), ("b", 1), ("b", 2)], names=["f", "s"]
-    )
-    ser = Series(range(4), index=idx)
-    result = ser.groupby(level=0).shift(1)
-    expected = ser.to_pandas().groupby(level=0).shift(1)
-    assert_eq(expected, result)
-
-
-@pytest.mark.parametrize(
-    "func", ["min", "max", "sum", "mean", "idxmin", "idxmax"]
-)
-@pytest.mark.parametrize(
-    "by,data",
-    [
-        ("a", {"a": [1, 2, 3]}),
-        (["a", "id"], {"id": [0, 0, 1], "a": [1, 2, 3]}),
-        ("a", {"a": [1, 2, 3], "b": ["A", "B", "C"]}),
-        ("id", {"id": [0, 0, 1], "a": [1, 2, 3], "b": ["A", "B", "C"]}),
-        (["b", "id"], {"id": [0, 0, 1], "b": ["A", "B", "C"]}),
-        ("b", {"b": ["A", "B", "C"]}),
-    ],
-)
-def test_group_by_reduce_numeric_only(by, data, func):
-    # Test that simple groupby reductions support numeric_only=True
-    df = cudf.DataFrame(data)
-    expected = getattr(df.to_pandas().groupby(by, sort=True), func)(
-        numeric_only=True
-    )
-    result = getattr(df.groupby(by, sort=True), func)(numeric_only=True)
-    assert_eq(expected, result)
-
-
-@pytest.mark.parametrize(
-    "op", ["cummax", "cummin", "cumprod", "cumsum", "mean", "median"]
-)
-def test_group_by_raises_string_error(op):
-    df = cudf.DataFrame({"a": [1, 2, 3, 4, 5], "b": ["a", "b", "c", "d", "e"]})
-
-    with pytest.raises(TypeError):
-        df.groupby(df.a).agg(op)
-
-
-@pytest.mark.parametrize(
-    "op",
-    [
-        "cummax",
-        "cummin",
-        "cumprod",
-        "cumsum",
-        "mean",
-        "median",
-        "prod",
-        "sum",
-        list,
-    ],
-)
-def test_group_by_raises_category_error(op):
-    df = cudf.DataFrame(
-        {
-            "a": [1, 2, 3, 4, 5],
-            "b": cudf.Series(["a", "b", "c", "d", "e"], dtype="category"),
-        }
-    )
-
-    with pytest.raises(TypeError):
-        df.groupby(df.a).agg(op)
-
-
-def test_ngroups():
-    pdf = pd.DataFrame({"a": [1, 1, 3], "b": range(3)})
-    gdf = cudf.DataFrame.from_pandas(pdf)
-
-    pgb = pdf.groupby("a")
-    ggb = gdf.groupby("a")
-    assert pgb.ngroups == ggb.ngroups
-    assert len(pgb) == len(ggb)
-
-
-def test_ndim():
-    pdf = pd.DataFrame({"a": [1, 1, 3], "b": range(3)})
-    gdf = cudf.DataFrame.from_pandas(pdf)
-
-    pgb = pdf.groupby("a")
-    ggb = gdf.groupby("a")
-    assert pgb.ndim == ggb.ndim
-
-    pser = pd.Series(range(3))
-    gser = cudf.Series.from_pandas(pser)
-    pgb = pser.groupby([0, 0, 1])
-    ggb = gser.groupby(cudf.Series([0, 0, 1]))
-    assert pgb.ndim == ggb.ndim
-
-
-@pytest.mark.skipif(
-    not PANDAS_GE_220, reason="pandas behavior applicable in >=2.2"
-)
-def test_get_group_list_like():
-    df = cudf.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6]})
-    result = df.groupby(["a"]).get_group((1,))
-    expected = df.to_pandas().groupby(["a"]).get_group((1,))
-    assert_eq(result, expected)
-
-    with pytest.raises(KeyError):
-        df.groupby(["a"]).get_group((1, 2))
-
-    with pytest.raises(KeyError):
-        df.groupby(["a"]).get_group([1])
-
-
-def test_get_group_list_like_len_2():
-    df = cudf.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6], "c": [3, 2, 1]})
-    result = df.groupby(["a", "b"]).get_group((1, 4))
-    expected = df.to_pandas().groupby(["a", "b"]).get_group((1, 4))
-    assert_eq(result, expected)
-
-
-def test_size_as_index_false():
-    df = pd.DataFrame({"a": [1, 2, 1], "b": [1, 2, 3]}, columns=["a", "b"])
-    expected = df.groupby("a", as_index=False).size()
-    result = cudf.from_pandas(df).groupby("a", as_index=False).size()
-    assert_groupby_results_equal(result, expected, as_index=False, by="a")
-
-
-def test_size_series_with_name():
-    ser = pd.Series(range(3), name="foo")
-    expected = ser.groupby(ser).size()
-    result = cudf.from_pandas(ser).groupby(ser).size()
-    assert_groupby_results_equal(result, expected)
-
-
-@pytest.mark.parametrize("op", ["cumsum", "cumprod", "cummin", "cummax"])
-def test_scan_int_null_pandas_compatible(op):
-    data = {"a": [1, 2, None, 3], "b": ["x"] * 4}
-    df_pd = pd.DataFrame(data)
-    df_cudf = cudf.DataFrame(data)
-    expected = getattr(df_pd.groupby("b")["a"], op)()
-    with cudf.option_context("mode.pandas_compatible", True):
-        result = getattr(df_cudf.groupby("b")["a"], op)()
-    assert_eq(result, expected)
-
-
-def test_agg_duplicate_aggs_pandas_compat_raises():
-    agg = {"b": ["mean", "mean"]}
-    dfgb = cudf.DataFrame({"a": [1, 1, 2], "b": [4, 5, 6]}).groupby(["a"])
-    with cudf.option_context("mode.pandas_compatible", True):
-        with pytest.raises(NotImplementedError):
-            dfgb.agg(agg)
-
-    with pytest.warns(UserWarning):
-        result = dfgb.agg(agg)
-    expected = cudf.DataFrame(
-        [4.5, 6.0],
-        index=cudf.Index([1, 2], name="a"),
-        columns=pd.MultiIndex.from_tuples([("b", "mean")]),
-    )
-    assert_groupby_results_equal(result, expected)
